requirements.txt
python-dotenv
langchain_google_genai
python-docx
pypandoc_binary
langchain_groq
langchain-ollama
fastapi[all]
reportlab
